added search optimization component, styling, and img with styling and added to app.js
