+++ horiseon-refactor/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import Header from './components/header';
+import Header from './components/header/header';
 import Img from './images/main.jpg';
+import Search from './components/searchOpt/search';
 import './styles/style.css';
 
 export default function App() {
@@ -8,6 +9,9 @@ export default function App() {
     <div>
     <Header />
     <img className="hero" src={Img} alt="digital marketing meeting img"></img>
+    <div className='main'>
+    <Search />
+    </div>
     </div>
   );
 }
