+++ horiseon-refactor/src/styles/style.css
@@ -45,3 +45,36 @@ h1 {
     background-size: cover;
     background-position: center;
 }
+p {
+    font-size: 16px;
+}
+.main{
+    display: flex;
+}
+.content-container{
+    flex: 2;
+}
+.content {
+    width: 75%;
+    display: inline-block;
+    margin-left: 20px;
+}
+.article {
+    margin-bottom: 20px;
+    padding: 50px;
+    height: 300px;
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+    background-color: #0072bb;
+    color: #ffffff;
+}
+#search-img{
+    max-height: 200px;
+}
+.article h2 {
+    margin-bottom: 20px;
+    font-size: 36px;
+}
+.float-left {
+    float: left;
+    margin-right: 25px;
+}
